+++ lib/index.ts
@@ -1,0 +1,2 @@
+export * from './core.js'
+export { AnalyticsProvider, useAnalytics } from './react/AnalyticsProvider.js'